memberroutine 수정

+++ src/main/java/com/bside/server/module/memberRoutine/repository/MemberRoutineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRoutineRepository extends JpaRepository<MemberRoutine, Long> {
+public interface MemberRoutineRepository extends JpaRepository<MemberRoutine, Integer> {
   Optional<MemberRoutine> findByMemberRoutineId(Integer MemberRoutineId);
 }
